add a test-headers make target and a make target for $(PCH) files
To test whether a header file successfully precompiles is a sanity
check that is stricter than merely testing whether the header file
parses. In other words, precompiling includes parsing and
additional steps that result in the intermediate form. The details
of precompilation (at least for clang) are explained at

http://clang.llvm.org/docs/PCHInternals.html

If foo.pch (with CC=clang++) or foo.gch (with CC=g++) is specified
in the call to make, then the makefile creates a precompiled
header from foo.hpp and redirects it to a bitbucket. If
test-headers is specified in the call to make, then every file
under src/stan that ends in .hpp is precompiled and redirected to
a bitbucket.