Use tab separation in ADVI logs
It can be useful to parse and plot the ELBO to demonstrate convergence. This is really tricky at the moment as spaces are used in as delim and also within field (notes in particular). I have yet to test this so take with massive dose of salt. I'm not sure if it prints neatly, never mind parses easier!